found a way to get username to auth every routes

+++ controllers/article.go
@@ -8,6 +8,7 @@ import (
 
 	"projet_wiki/database"
 	"projet_wiki/models"
+	"projet_wiki/controllers/auth"
 
 	"github.com/gorilla/mux"
 )
@@ -104,6 +105,26 @@ func DeleteArticle(w http.ResponseWriter, r *http.Request) {
 	if err != nil {
 		fmt.Println(err)
 	}
+
+	UserAuth, errAuth := auth.GetRequestUser(r)
+	if errAuth != nil {
+		fmt.Println("user not connected")
+		return
+	}
+
+	UserId := article.UserID
+	var user models.User
+	db.Where("id = ?", UserId).Find(&user)
+	
+	if err != nil {
+		fmt.Println(err)
+	}
+
+	if UserAuth.Username != user.Username {
+		fmt.Println("Incorrect account rights ")
+	    return
+	}
+
 	db.Delete(&article)
 }
 
@@ -121,6 +142,25 @@ func UpdateArticle(w http.ResponseWriter, r *http.Request) {
 	err := db.Where("id = ?", ID).Find(&article)
 	if err != nil {
 		fmt.Println(err)
+	}
+
+	UserAuth, errAuth := auth.GetRequestUser(r)
+	if errAuth != nil {
+		fmt.Println("user not connected")
+		return
+	}
+
+	UserId := article.UserID
+	var user models.User
+	db.Where("id = ?", UserId).Find(&user)
+	
+	if err != nil {
+		fmt.Println(err)
+	}
+
+	if UserAuth.Username != user.Username {
+		fmt.Println("Incorrect account rights ")
+	    return
 	}
 
 	article.Title = title
