+++ controllers/comment.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"net/http"
 	"strconv"
-
+	"projet_wiki/controllers/auth"
 	"projet_wiki/database"
 	"projet_wiki/models"
 
@@ -94,6 +94,26 @@ func DeleteComment(w http.ResponseWriter, r *http.Request) {
 	if err != nil {
 		fmt.Println(err)
 	}
+
+	UserAuth, errAuth := auth.GetRequestUser(r)
+	if errAuth != nil {
+		fmt.Println("user not connected")
+		return
+	}
+
+	UserId := comment.UserID
+	var user models.User
+	db.Where("id = ?", UserId).Find(&user)
+	
+	if err != nil {
+		fmt.Println(err)
+	}
+
+	if UserAuth.Username != user.Username {
+		fmt.Println("Incorrect account rights ")
+	    return
+	}
+
 	db.Delete(&comment)
 }
 
@@ -110,6 +130,25 @@ func UpdateComment(w http.ResponseWriter, r *http.Request) {
 	err := db.Where("id = ?", ID).Find(&comment)
 	if err != nil {
 		fmt.Println(err)
+	}
+
+	UserAuth, errAuth := auth.GetRequestUser(r)
+	if errAuth != nil {
+		fmt.Println("user not connected")
+		return
+	}
+
+	UserId := comment.UserID
+	var user models.User
+	db.Where("id = ?", UserId).Find(&user)
+	
+	if err != nil {
+		fmt.Println(err)
+	}
+
+	if UserAuth.Username != user.Username {
+		fmt.Println("Incorrect account rights ")
+	    return
 	}
 
 	comment.Content = content
